LayoutTests/http/wpt/web-animations/interfaces/KeyframeEffect/setKeyframes-expected.txt


PASS Keyframes can be replaced with an empty keyframe 
PASS Keyframes can be replaced with a one property two value property-indexed keyframes specification 
FAIL Keyframes can be replaced with a one shorthand property two value property-indexed keyframes specification assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,offset" but got "computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
FAIL Keyframes can be replaced with a two property (one shorthand and one of its longhand components) two value property-indexed keyframes specification assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,marginTop,offset" but got "computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
PASS Keyframes can be replaced with a two property two value property-indexed keyframes specification 
PASS Keyframes can be replaced with a two property property-indexed keyframes specification with different numbers of values 
PASS Keyframes can be replaced with a property-indexed keyframes specification with an invalid value 
PASS Keyframes can be replaced with a one property two value property-indexed keyframes specification that needs to stringify its values 
FAIL Keyframes can be replaced with a property-indexed keyframes specification with a CSS variable reference assert_equals: value for 'left' on ComputedKeyframe #0 expected "var(--dist)" but got "auto"
FAIL Keyframes can be replaced with a property-indexed keyframes specification with a CSS variable reference in a shorthand property assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,offset" but got "computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
PASS Keyframes can be replaced with a one property one value property-indexed keyframes specification 
PASS Keyframes can be replaced with a one property one non-array value property-indexed keyframes specification 
PASS Keyframes can be replaced with a one property two value property-indexed keyframes specification where the first value is invalid 
PASS Keyframes can be replaced with a one property two value property-indexed keyframes specification where the second value is invalid 
PASS Keyframes can be replaced with a one property one keyframe sequence 
PASS Keyframes can be replaced with a one property two keyframe sequence 
PASS Keyframes can be replaced with a two property two keyframe sequence 
FAIL Keyframes can be replaced with a one shorthand property two keyframe sequence assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,offset" but got "computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
FAIL Keyframes can be replaced with a two property (a shorthand and one of its component longhands) two keyframe sequence assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,marginTop,offset" but got "computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
FAIL Keyframes can be replaced with a keyframe sequence with duplicate values for a given interior offset Type error
FAIL Keyframes can be replaced with a keyframe sequence with duplicate values for offsets 0 and 1 Type error
FAIL Keyframes can be replaced with a two property four keyframe sequence Type error
PASS Keyframes can be replaced with a single keyframe sequence with omitted offset 
PASS Keyframes can be replaced with a single keyframe sequence with null offset 
PASS Keyframes can be replaced with a single keyframe sequence with string offset 
PASS Keyframes can be replaced with a one property keyframe sequence with some omitted offsets 
PASS Keyframes can be replaced with a one property keyframe sequence with some null offsets 
PASS Keyframes can be replaced with a two property keyframe sequence with some omitted offsets 
PASS Keyframes can be replaced with a one property keyframe sequence with all omitted offsets 
FAIL Keyframes can be replaced with a keyframe sequence with different easing values, but the same easing value for a given offset Type error
FAIL Keyframes can be replaced with a keyframe sequence with different composite values, but the same composite value for a given offset Type error
PASS Keyframes can be replaced with a one property two keyframe sequence that needs to stringify its values 
FAIL Keyframes can be replaced with a keyframe sequence with a CSS variable reference assert_equals: value for 'left' on ComputedKeyframe #0 expected "var(--dist)" but got "auto"
FAIL Keyframes can be replaced with a keyframe sequence with a CSS variable reference in a shorthand property assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,offset" but got "computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
FAIL Keyframes can be replaced with a keyframe sequence where shorthand precedes longhand assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,marginRight,offset" but got "computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
FAIL Keyframes can be replaced with a keyframe sequence where longhand precedes shorthand assert_equals: properties on ComputedKeyframe #0 expected "computedOffset,easing,margin,marginRight,offset" but got "computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset"
FAIL Keyframes can be replaced with a keyframe sequence where lesser shorthand precedes greater shorthand assert_equals: properties on ComputedKeyframe #0 expected "border,borderLeft,computedOffset,easing,offset" but got "borderBottomColor,borderBottomStyle,borderBottomWidth,borderImageOutset,borderImageRepeat,borderImageSlice,borderImageSource,borderImageWidth,borderLeftColor,borderLeftStyle,borderLeftWidth,borderRightColor,borderRightStyle,borderRightWidth,borderTopColor,borderTopStyle,borderTopWidth,computedOffset,easing,offset"
FAIL Keyframes can be replaced with a keyframe sequence where greater shorthand precedes lesser shorthand assert_equals: properties on ComputedKeyframe #0 expected "border,borderLeft,computedOffset,easing,offset" but got "borderBottomColor,borderBottomStyle,borderBottomWidth,borderImageOutset,borderImageRepeat,borderImageSlice,borderImageSource,borderImageWidth,borderLeftColor,borderLeftStyle,borderLeftWidth,borderRightColor,borderRightStyle,borderRightWidth,borderTopColor,borderTopStyle,borderTopWidth,computedOffset,easing,offset"
PASS Keyframes can be replaced with a two property keyframe sequence where one property is missing from the first keyframe 
PASS Keyframes can be replaced with a two property keyframe sequence where one property is missing from the last keyframe 
FAIL Keyframes can be replaced with a keyframe sequence with repeated values at offset 1 with different easings Type error
PASS KeyframeEffect constructor throws with keyframes with an out-of-bounded positive offset 
PASS KeyframeEffect constructor throws with keyframes with an out-of-bounded negative offset 
PASS KeyframeEffect constructor throws with keyframes not loosely sorted by offset 
PASS KeyframeEffect constructor throws with property-indexed keyframes with an invalid easing value 
PASS KeyframeEffect constructor throws with a keyframe sequence with an invalid easing value 
PASS KeyframeEffect constructor throws with keyframes with an invalid composite value 
